refactor: Refactored controller to use  ResponseEntity  for HTTP responses, standardizing status codes and response bodies.

+++ src/main/java/dev/wesleyjunior/RegisterNinjas/Ninjas/NinjaController.java
@@ -1,5 +1,7 @@
 package dev.wesleyjunior.RegisterNinjas.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,29 +19,52 @@ public class NinjaController {
     }
 
     @GetMapping("/list")
-    public List<NinjaDTO> listNinjas() {
-        return ninjaService.listNinjas();
+    public ResponseEntity<List<NinjaDTO>> listNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     @GetMapping("/list/{uuid}")
-    public NinjaDTO listNinjasById(@PathVariable UUID uuid) {
-        return ninjaService.listNinjasById(uuid);
+    public ResponseEntity<?> listNinjasById(@PathVariable UUID uuid) {
+        NinjaDTO ninjaDTO = ninjaService.listNinjasById(uuid);
+        if (ninjaDTO != null) {
+            return ResponseEntity.ok(ninjaDTO);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja id:" + uuid + " not found.");
     }
 
+
     @PostMapping("/create")
-    public NinjaDTO createNinja(@RequestBody NinjaDTO ninja) {
-        return ninjaService.createNinja(ninja);
+    public ResponseEntity<String> createNinja(@RequestBody NinjaDTO ninja) {
+        NinjaDTO ninjaDTO = ninjaService.createNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja created successfully." + "Ninja's name: " + ninjaDTO.getName());
     }
 
     @DeleteMapping("/delete/{uuid}")
-    public void deleteNinjas(@PathVariable UUID uuid) {
-        ninjaService.deleteNinjaById(uuid);
+    public ResponseEntity<String> deleteNinjas(@PathVariable UUID uuid) {
+        NinjaDTO ninjaDTO = ninjaService.listNinjasById(uuid);
+
+        if (ninjaDTO != null) {
+            ninjaService.deleteNinjaById(uuid);
+            return ResponseEntity.ok("Ninja deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja not found.");
+        }
     }
 
     @PutMapping("/edit/{uuid}")
-    public NinjaDTO editNinjas(@PathVariable UUID uuid, @RequestBody NinjaDTO ninjaUpdated) {
+    public ResponseEntity<String> editNinjas(@PathVariable UUID uuid, @RequestBody NinjaDTO ninjaUpdated) {
+        NinjaDTO ninjaDTO = ninjaService.listNinjasById(uuid);
 
-        return ninjaService.updateNinja(ninjaUpdated, uuid);
+        if (ninjaDTO != null) {
+            ninjaService.updateNinja(ninjaUpdated, uuid);
+            return ResponseEntity.ok("Ninja updated successfully. Ninja updated: " + ninjaUpdated);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja not found.");
+        }
 
     }
